+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add buttons to add CoD mage skills to the actor magic tab
 - Add source field to CoD spell forms
 - Make search work on the inventory tab as well
+- Mods when fix skill is 1 or more
+  - inventory tab will have an installed mod location and show how many you can do
+  - stats tab is where to set the max
+  - items can be set as a mod which will give a pot to specify he location that it is installed, which will show as a badge on the item list
 
 ### Fixed